requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
bleach==6.1.0
Pillow==10.1.0
python-magic==0.4.27
Flask-Limiter==3.5.0
redis==5.0.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
gunicorn==21.2.0
